Moved/renamed files in sdk and calibration app.
