+++ calibration-app/app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-	package="org.dpppt.android.calibration">
+	package="org.coralibre.android.calibration">
 
 	<application
-		android:name="org.dpppt.android.calibration.MainApplication"
+		android:name="org.coralibre.android.calibration.MainApplication"
 		android:allowBackup="false"
 		android:icon="@mipmap/ic_launcher"
 		android:label="@string/app_name"
@@ -21,7 +21,7 @@
 		android:networkSecurityConfig="@xml/network_security_config"
 		android:supportsRtl="true"
 		android:theme="@style/AppTheme">
-		<activity android:name="org.dpppt.android.calibration.MainActivity">
+		<activity android:name="org.coralibre.android.calibration.MainActivity">
 			<intent-filter>
 				<action android:name="android.intent.action.MAIN" />
 				<category android:name="android.intent.category.LAUNCHER" />
