+++ calibration-app/app/src/main/java/org/coralibre/android/calibration/util/OnTextChangedListener.java
@@ -8,7 +8,7 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-package org.dpppt.android.calibration.util;
+package org.coralibre.android.calibration.util;
 
 import android.text.Editable;
 import android.text.TextWatcher;
