+++ calibration-app/settings.gradle
@@ -10,4 +10,4 @@
 
 rootProject.name = 'DP3T Calibration App'
 include ':app', ':dp3t-sdk'
-project(':dp3t-sdk').projectDir = new File('../dp3t-sdk/sdk')
+project(':dp3t-sdk').projectDir = new File('../coralibre-sdk/sdk/')
