+++ coralibre-sdk/sdk/src/calibration/java/org/coralibre/android/sdk/internal/logger/LogEntry.java
