+++ coralibre-sdk/sdk/src/calibration/java/org/coralibre/android/sdk/util/DeviceHelper.java
